refactor(lotto): rank 패키지 내 위치변경
