+++ src/test/kotlin/lotto/domain/LottoBunchTest.kt
@@ -1,6 +1,5 @@
 package lotto.domain
 
-import lotto.constant.Rank
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
